pyproject.toml
[project]
name = "llm-router"
version = "0.1.0"
description = "LLM Routing Service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "PyYAML>=6.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "openai>=1.54.0",
    "google-generativeai>=0.8.0",
    "tiktoken>=0.8.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-dotenv>=0.5.0",
    "pytest-cov>=6.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
pythonpath = [
    ".",
]
testpaths = [
    "tests",
]
asyncio_mode = "auto"
env_files = [
    ".env.test",
]
env_override_existing_values = 1
